Bump swagger cli / golangci-lint / UI related package.json

+++ swagger_gen/restapi/operations/constraint/delete_constraint_parameters.go
@@ -16,7 +16,8 @@ import (
 )
 
 // NewDeleteConstraintParams creates a new DeleteConstraintParams object
-// no default values defined in spec.
+//
+// There are no default values defined in the spec.
 func NewDeleteConstraintParams() DeleteConstraintParams {
 
 	return DeleteConstraintParams{}
@@ -74,7 +75,6 @@ func (o *DeleteConstraintParams) BindRequest(r *http.Request, route *middleware.
 	if err := o.bindSegmentID(rSegmentID, rhkSegmentID, route.Formats); err != nil {
 		res = append(res, err)
 	}
-
 	if len(res) > 0 {
 		return errors.CompositeValidationError(res...)
 	}
@@ -107,7 +107,7 @@ func (o *DeleteConstraintParams) bindConstraintID(rawData []string, hasKey bool,
 // validateConstraintID carries on validations for parameter ConstraintID
 func (o *DeleteConstraintParams) validateConstraintID(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("constraintID", "path", int64(o.ConstraintID), 1, false); err != nil {
+	if err := validate.MinimumInt("constraintID", "path", o.ConstraintID, 1, false); err != nil {
 		return err
 	}
 
@@ -140,7 +140,7 @@ func (o *DeleteConstraintParams) bindFlagID(rawData []string, hasKey bool, forma
 // validateFlagID carries on validations for parameter FlagID
 func (o *DeleteConstraintParams) validateFlagID(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("flagID", "path", int64(o.FlagID), 1, false); err != nil {
+	if err := validate.MinimumInt("flagID", "path", o.FlagID, 1, false); err != nil {
 		return err
 	}
 
@@ -173,7 +173,7 @@ func (o *DeleteConstraintParams) bindSegmentID(rawData []string, hasKey bool, fo
 // validateSegmentID carries on validations for parameter SegmentID
 func (o *DeleteConstraintParams) validateSegmentID(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("segmentID", "path", int64(o.SegmentID), 1, false); err != nil {
+	if err := validate.MinimumInt("segmentID", "path", o.SegmentID, 1, false); err != nil {
 		return err
 	}
